Enable tempest-full on stable/pike
Add stable/pike on tempest-full.
The job will be voting on stable/pike so we need bitrot
jobs on Tempest master as well. Not the py3 since it was
not running on Pike yet.

Once this lands we can add bitrot jobs to d-g as well,
and filter the legacy job out of pike in the integrated
templates.

Depends-on: https://review.openstack.org/545258/
Change-Id: I1ac0e54eab22786575e11f4d34b7e15e643038c4