Remove duplicate unit test module
For some reason the test_object_client unit test was never removed after
it was copied to the lib unit tests, as part of the lib migration. I'm
not sure why this isn't failing because this module has a broken import,
and this is a bit concerning that nothing caught this error. But while
we investigate that let's at least remove the dead code.

Change-Id: I9d46067f325eac858ee9850960b9dde8cc34e763