Cleanup selection list args in tempest run
The selection list arguments had a mutually exclusive flag set for no
reason. There previously was a limitation in how the flags were used
previously which made those flags mutually exclusive, however that
hasn't been the case for some time. At the same time the docs on
selection needed to be updated, because they showed incorrect internal
details (even prior to the stestr migration) for the blacklist file.
Instead of trying to detail the entire explanation, it now just links
to the stestr docs.

Change-Id: I90dfc15765057f876ebaf66abf5af24de0f675b9