Run all scenario in parallel
We stopped running all scenarios in parallel,
which leads to longer job times and less concurrent test run.

The full parallel run had too high failure rate,
this job is supposed to prove it is possible to return back to
the old way. If the issue(s) are still exists it will
produce usable artifacts for further debugging.

Change-Id: I776ebb7761d90d2490cf49f9df9b322c2b34da59