Add nova-cells-v1 to experimental queue jobs
The legacy-tempest-dsvm-cells job is being moved into the
nova repo and renamed to nova-cells-v1. This change adds
the new job name to the in-tree definition of the experimental
queue jobs that run on tempest changes. A project-config change
will depend on this patch to undefine the legacy job name from
being used on tempest changes.

Depends-On: https://review.openstack.org/549780
Change-Id: If6b7e59e71739361dcb0d246fca3fcf08af3514d