Switch to stestr in README.rst and the method name
This commit switches to stestr in README.rst and the method name. We
already switched to use stestr instead of testr/ostestr. However,
a few are still remaining in README.rst and the unit test name. This
patch leaves some `testr` things because they are for the 'legacy run
method'. So, we don't need to remove them completely.

Change-Id: I0e8e19e2150d6e0ce9a299a10bdd51652d86497e