Don't use hypervisor_hostname as compute hostname
Several tempest tests are failing when using the VMware virt driver
where compute hostname != hypervisor_hostname. Those are different
things and cannot be used interchangeably. This patch fixes a
regression introduced with commit 1dc2e231 where the hosts_client was
replaced with hypervisor_client in order to select hosts which are both
up and enabled.

This patch changes the implementation of get_host_other_than() to use
services_client which still allows to select nova computes which are up
and enabled but also gives us the compute hostname.

Change-Id: Ie6df1192e1ca83977dccc8e6b9331dfe874c8aba
Closes-Bug: #1759801