Add common method to check the microversion compatibility
As there are lot of API microversion with behavior change,
tests internal checks needs to be adjusted according to
requested microversion.
Example: https://review.openstack.org/#/c/557280/

That need to compare the requested microversion with max supported
microversion. Same comparison is needed multiple places in multiple
files.

This commit adds a common method to do the microversion comparison
which can be used in many place.

Change-Id: Ib54777e5ce101bab6c2980ae69ef2c5d978ff064