Add nova-live-migration to experimental queue jobs
The legacy-tempest-dsvm-multinode-live-migration job is being moved
into the nova repo and renamed to nova-live-migration. This change adds
the new job name to the in-tree definition of the experimental queue
jobs that run on tempest changes. A project-config change will depend
on this patch to undefine the legacy job name from being used on
tempest changes.

Depends-On: https://review.openstack.org/555945

Change-Id: If8150527a358cfe634a604259af47419369fbdc3