Run nova-live-migration job in check as n-v
nova define the 'nova-live-migration' job in-tree and run
this job to nova check and gate pipeline as voting job
- https://github.com/openstack/nova/blob/master/.zuul.yaml

Tempest has that job running as experimental.
Live migration tests are running as part of
'legacy-tempest-dsvm-neutron-scenario-multinode-lvm-multibackend' job
as non voting check pipeline job.

Idea is:
- Run live migration tests via 'nova-live-migration' job which has much better
  set up of live migration.
- After that make that legacy scenario job as scenario tests only and
  remove running the live migration tests in scenario job.

That will give better clarity on which job run which tests.
As live migration test are running as voting job, Tempest also should
test them in same way.

This commit move nova-live-migration job to check pipeline.

There are frequesnt failure on this job[1] so let's keep this
n-v as of now and later we want to make it voting which will be
same as nova gate.

[1] http://logstash.openstack.org/#/dashboard/file/logstash.json?query=message:%5C%22Live%20Migration%20failed.%20Migrations%20list%20for%20Instance%20%5C%22

Change-Id: Ia4d13ae6dd8bcbfd40dcd624e1837cc07898bb36