Remove test_get_service_by_service_and_host_name
By default, devstack runs in superconductor mode, meaning there
are 2 conductor services, one is in the cell0 database and one
is in the cell1 database. These get reported with the same service
binary "nova-conductor" and on the same host (since it's single
node devstack). This can cause failures in this test which
randomly picks a service from the list and then does a query by
hostname and binary, and if it picks nova-conductor, it can fail
because it's expecting one but there can be two conductors in this
type of deployment.

This test shouldn't even exist in tempest since it can be (and is)
tested in the nova tree and it's not an interop guideline since
it's an admin-only API by default.

Change-Id: I40a776959768f4422dd7960680a877db96cc3b9b
Closes-Bug: #1771817