Use inline literal markup for options in releasenotes
This commit makes the releasenotes to use 'inline literal markup[0]
because it would be accurate rendering for commandline
options. Otherwise, dashes are changed into an en/em dash entity. And
this commit also fixes some inline literal markups which I
noticed. Ideally, we should fix like that things in one
patch. However, it's a bit hard to find them all..

[0] http://www.sphinx-doc.org/en/master/usage/restructuredtext/basics.html#roles

Change-Id: I8f7691b3ff40ccde051620f44217dab244d76960