Cap nova-network tests at Queens
Nova is working on removing the nova-network only
compute REST APIs in the Rocky release, which means
any requests to those APIs after Queens will result
in a 410 response, regardless of microversion or
whether or not neutron is being used.

The os-virtual-interfaces API is being removed in
nova change: https://review.openstack.org/569923/

As a result, os-virtual-interfaces tests will no longer
works against Rocky+ versions of nova.

This adds a new check in the base compute API test
class to determine, based on the max compute API
microversion, which release is being tested and if
Rocky+ the nova-network specific tests will be skipped.

Subsequent nova-network specific tests will be skipped
as the API removal changes are proposed in nova.

Related to blueprint remove-nova-network

Change-Id: I7aea99d7c1818b8edcda41ba4eaa062a8ea069eb