Fix test_volume_swap_with_multiattach
Nova change https://review.openstack.org/572790/ is fixing
a bug such that the compute API would allow people to
swap from a multiattach volume with multiple read/write
attachments, which could lead to data loss if the
secondary attachment is writing to the source volume while
it's being copied to the target volume.

As a result, test_volume_swap_with_multiattach needs to be
changed such that the volume we're swapping from has only
read-only attachments.

Change-Id: Ida387c600016b451e01118bc2c76662b46670288
Related-Bug: #1775418