Move nova cell v1 job to check pipeline
Nova cell v1 job run as experimental and it is
obvious that we miss testing the tempest changes against
this job which can leads to gate block on nova side.
One example is https://bugs.launchpad.net/nova/+bug/1776684

nova cell v1 is deprecated but not yet removed and this job
run on nova check and gate pipeline.

This commit moves the cell v1 job from experimental pipeline
to check pipeline on tempest side so that we can monitor it.

Once nova wil drop cell v1, Tempest can drop this job.

Change-Id: Ic01e867d590810f620b141a554a596ddd5e73fed