Rename scenario job and make it voting
"legacy-tempest-dsvm-neutron-scenario-multinode-lvm-multibackend" job
which was named as "tempest-scenario-multinode-lvm-multibackend" run
all scenario tests including slow tests.

This job has been added as n-v to check the stability of slow tests and
it has been passing consistently after that.

Let's make this job as voting and run on check and gate pipeline so that
all scenario test can be monitored consistently.

For clear name, this commit also rename this job to "tempest-scenario-all".

Change-Id: I3522a374bcae1f285d464386d7466912f613e06a