Cleanup v18.0.0 release notes' prelude
The v18.0.0 tempest release notes had cluttered the prelude section with
random details about a OVS feature. This did not belong in the prelude
because this section is for high level details about the release. This
normally just is things like adds support for release X, or removal of
unsupported release W. This commit attempts to clean this up by moving
the comments about the rationale for a flag to the feature release
notes for the new flag.

Change-Id: I4831688b55d1e106880b0e4e6998c407eb60fc17