Move volume v2 service client unit tests to v3 dir
This patchset builds off work done recently in [0] to move the
v2 service clients to v3 dir -- this begins moving the
v2 service client unit tests to v3 dir.

This is done for the following reasons:

* v2 is deprecated and subsequently really just calls the
  v3 clients referentially [1]
* unnecessary deprecation warnings are emitted for unit
  tests [2], they should be cleaned up
* it is much harder to figure out what coverage is missing:
  you need to look at v2 and v3 dir together to figure it
  out
* poor code organization: it is much cleaner to consolidate
  the tests together

This patchset only does this for volumes v2 client -- if
there is agreement we can proceed with the other client
tests next.

[0] https://review.openstack.org/#/c/536121/
[1] e.g. https://github.com/openstack/tempest/blob/2b6e955c59d814538ceeb58d9160433eb915688d/tempest/lib/services/volume/v2/volumes_client.py#L17
[2] e.g. http://logs.openstack.org/65/578765/2/check/openstack-tox-py35/88f9858/job-output.txt.gz#_2018-06-28_13_43_16_409494

Change-Id: I3b4715d0a667ddc4e187b9353b68dd579aefa2c5