refactor: Combine nearly duplicate group tests together
This patchset combines test_group_update_with_few_fields with
test_group_create_update_get since the first one is almost the
exact same as the update scenario in the second test.

test_group_update_with_few_fields just updates 1 field and
checks whether the other fields remain unchanged. This is a
unique test case that is very niche and so shouldn't really be
its own test case. It can be combined with
test_group_create_update_get.

Also this patchset adds better comments and also adds assertions
for checking that the domain_id of the group is correct.

Change-Id: I4121b85aa271ccd7df30d7675bd923cb84a5673a