docs: Add information on test removal/interop to REVIEWING
This patch set adds information on test removal, relocation,
renaming to REVIEWING because it is important that such
actions do not break interop. Interop is closely tied to
Tempest because it directly references Tempest tests that
are not only expected to exist, but to also work.

The same is true of breaking blacklist/whitelist
references to Tempest tests, this is also included
in the new documentation section.

It is important that there be REVIEWING guidelines
in place to assist reviewers understand this importance.
Also references are included for defcore/interop to help
users find more information on these topics.

Currently interop is only mentioned in 1 place in Tempest [0]
and yet there is little information about it. This patchset
aims to make it easier to find more information about it
for reviewers and users alike.

[0] http://git.openstack.org/cgit/openstack/tempest/tree/doc/source/test_removal.rst

Change-Id: Ifbde674b42355077fcd8daa07be8be1248b77b0f