Check API extension 'filter-validation'
Neutron is planing to employ a strict validation
against filter parameters [1]. The motivation is to align with the
OpenStack API guideline [2]. In particular, the API guideline
mentioned the following:

"... if the API supports query parameters and a request
contains an unknown or unsupported parameter, the server should
return a 400 Bad Request response. Invalid values in the request URL
should never be silently ignored, as the response may not match the
client’s expectation..."

After discussion with the API work group [3], the neutron team decides
to make a backward-incompatible change (with a config option to
rollback to the old API behavior).

This patch checks if 'filter-validation' is presented and assert
the old API behavior (silently ignore invalid parameter) if the
extension is not presented.

[1] https://bugs.launchpad.net/neutron/+bug/1749820
[2] http://specs.openstack.org/openstack/api-wg/guidelines/http/
    response-codes.html#failure-code-clarifications
[3] http://eavesdrop.openstack.org/meetings/api_sig/2018/
    api_sig.2018-03-15-16.00.log.html#l-120

Change-Id: Ia6881832bf686525b018613af91e683e6f0b2470