Support immutable user source
If the keystone user source is immutable, such as an LDAP active
directory implementation, tempest tests that try to create, delete
or modify a user will fail. Instead of failing, we would like them to
skip. Proposed is an additional config setting in the auth group
called immutable_user_source that is defaulted to false.

To handle this on a test by test basis and avoid modifying behavior
that depends on the identity base class, we propose using a
testtools decorator to skip tests based on the new config setting.

One test class so far has been adjusted to use the decorator, and if
the config setting is set to true, the tests will skip accordingly.

Co-Authored-By: Michael Beaver <michaelbeaver64@gmail.com>
Partial-Bug: 1777047
Change-Id: Idc09d6272386f026a899787c2151745916a7228d