Add test for create volume from snap without size
If a user creates a volume from a snapshot but does not specify a size,
Cinder will default to creating a volume with the same original size
from when the snapshot was created. We had test coverage for the case of
specifying a larger volume size, but not for this default size. This
adds a single test to verify things work as expected without passing in
a value for the new volume size.

Change-Id: If1704af4c79281df2fc05d8198b30a2653f53d74
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>