Immutable user source: test_default_project_id
If the keystone user source is immutable, such as an LDAP
active directory implementation, tempest tests that try
to create or delete a user will fail. Instead of failing,
we would like them to skip. This change uses a testtools
decorator to avoid unnecessary modifications and allow those tests
to skip. In [1], I introduced the config setting that allows
this to happen.

[1] https://review.openstack.org/#/c/585536/

Change-Id: I4d928fdbf04de65a73977dbe4a818dff4cb2aa8d
Partial-Bug: 1777047