Add server schema for microversion 2.8
This commit adds servers schema for microversion 2.8 [0]:

Add ‘mks’ protocol and ‘webmks’ type for remote consoles.

The appropriate api_schema files are updated with the
2.8 microversion changes.

Note that no test is added to cover this new microversion
since it isn't possible to test mks protocol in Tempest
gates so as per REVIEWING documentation [1] we shouldn't
add such a test. So, we could only cover the vnc protocol
for 2.8 but this would also not be of much value since
the vnc value (instead of mks) would be returned by the
server.

[0] https://docs.openstack.org/nova/latest/reference/api-microversion-history.html#id7
[1] https://docs.openstack.org/tempest/latest/REVIEWING.html#ensuring-code-is-executed

partially-implements: blueprint full-schema-for-all-microversions
Change-Id: I5a6d2852db8e2a67013eee71a094e9b521a92524