Fill quota schema for microversion 2.36/2.57
Network related attributes are removed from quota set from
microversion 2.36, and injected file related attributes are
removed from microversion 2.57, so this is to remove them
from schema of quota set according to the microversion.

Because show_quota_set uses schema get_quota_set when detail
is False and uses schema get_quota_set_details when detail
is True, so this is also to change one get_quota_set calling
to be with detail=True, so to cover the schema check of
get_quota_set_details.

https://developer.openstack.org/api-ref/compute/#show-a-quota

Change-Id: Ib3762ac10d9fef42e4747710b35073797a36636a
partially-implements: blueprint full-schema-for-all-microversions