Delete snapshot in test_image_defined_boot_from_volume
We've observed error messages in the gate during this test:

   Failed to delete volume: 16f9252c-4036-463b-a053-60d4f46796c1 due to
   Invalid input received: Invalid volume: Volume status must be
   available or error or error_restoring or error_extending and must
   not be migrating, attached, belong to a consistency group or have
   snapshots. (HTTP 400)

Upon investigating, I found that it's the deletion of the first server
in the test that fails for "Invalid volume" and it's because we delete
the server *after* we've created a snapshot of it (which creates a
volume snapshot). So the first volume, backing the first server, has a
snapshot and that first volume cannot be deleted until the snapshot is
deleted.

This adds a deletion of the volume snapshot and moves the deletion of
the first server after it.

Related-Bug: #1527623

Change-Id: Ia3ccbf8c11d3408742a1f75b1795caa2cc7bbe64