Do not pass device when calling compute.base.attach_volume
As of the 12.0.0 Liberty release, the Nova libvirt driver
no longer honors a user-supplied device name, so better not
pass device when calling compute.base.attach_volume.
https://developer.openstack.org/api-ref/compute/#attach-a-volume-to-an-instance

NOTE: We still keep passing device in test_attach_detach_volume,
because it will ssh to vm to check the device name, in that case,
CONF.compute.volume_device_name must be set the equal value as
the libvirt auto-assigned one. Whether to remove that check
is out of the range of this patch.

Change-Id: I4483836abedab2550731fef5b2625282d7a2298d