Interface tests: wait for server to become SSHable
Attaching and detaching a network interface depends on guest OS
cooperation, as it needs to properly handle the ACPI event. Some
hardware/hypervisor configurations make the guest run slow enough that
it doesn't process the events in time, leading to the interface
attachment tests to fail unpredictably. This patch uses
validate_authentication() to wait for the server to become SSHable. By
doing so, we ensure that it'll properly process the ACPI event(s).

Change-Id: I2e7fde779328664ac1fc91eef6ab702dd5321f84