Add missing test for "Show API v2 details" action
Add test to cover the request to "Show API v2 details"[0] action.
Add the show_version function to the NetworkVersionsClient class.
Modify the list_versions function of the NetworkVersionsClient class
to use the .get() function instead of the .raw_request() function
because using the .get() makes the function simplier and abstracts out
the logging and retrying a request for us.
Add the unit tests for the show_version function.

[0] https://developer.openstack.org/api-ref/network/v2/index.html#show-api-v2-details

Change-Id: I7ebea0c53437e929fc613495bcb8a8f8838c2044