Update REVIEWING docs with fast-track approval process
This patch set follows through on discussion related to fast
tracking changes in Tempest that are required to unblock other
project gates. So, the "When to approve" section of the REVIEWING
documentation has been expanded to include a fast track
provision: that 1 core reviwer can +2 a change and approve it,
provided that the core reviwer belonging to the project with
the failing gate +1's the change, too.

Change-Id: Iafb9de650cd79b3c72b03e20690dd80a95308b5b