Mark some slow tests as slow so they move out of the main job
We've seen a trend of tempest-full jobs timing out because of
tests taking a long time to complete which are not marked slow:

  http://status.openstack.org/elastic-recheck/#1783405

Based on some analysis of the slowest tests:

  https://ethercalc.openstack.org/dorupfz6s9qt

This change marks several tests as slow and gives reasons why
they are marked as slow - not only because they are slow, but
also because they are less common scenarios. These will still
be run in the tempest-slow job so we aren't losing test coverage
with this change.

If this makes the tempest-slow job itself take significantly
longer, we should consider running the tempest.api* slow tests
concurrently (like they are in tempest-full) and continue to
run the tempest.scenario* tests in serial. The tempest-slow job
currently runs all slow tests in serial, so this will make the
job take longer.

Partial-Bug: #1783405

Change-Id: I2a0e154ba38c7407b41b7e986a0bf9b451c65cae