swift: container ACL - use ID's for tenant and user
Switch to using ID's instead of names for the test tenant and user.

Under Keystone v3 names (vs ids) are only supported for members of
the 'default' domain, so if accounts are created in a different domain
the container ACL tests will fail.

Switch to using ID's which are guaranteed to be unique across domains,
resolving test failures in this type of deployment configuration.

Change-Id: I79e5754379e3cf72d2a9858e6e69e8c190b0b2ed