Wait for IP add/remove in test_add_remove_fixed_ip
Adding/removing fixed IPs to/from a server is an
asynchronous operation. The test_add_remove_fixed_ip test
was not waiting for the fixed IP to be added to/removed from
the server, which means that it's possible for the server to
be getting deleted during tearDown before the removeFixedIp
operation has completed and will end up failing in the
nova-compute service. That failure wouldn't cause the test
to fail, however, since the test wasn't waiting for the
operation to complete at all.

This adds code to wait for the added fixed IP to show up
on the server and for the removed fixed IP to be removed
from the server's list of addresses. Furthermore, the test
is changed to specifically remove the fixed IP that was
added.

Change-Id: I10a34fb12b595400fc15d935070f130c8e3d3963
Closes-Bug: #1794558