GABBI_TEMPEST_PATH in tox passenv and devstack playbook
The gabbi-tempest plugin uses an environment variable,
GABBI_TEMPEST_PATH, to identify directories in which to
find the gabbi [1] tests that will be run. This will be
used by a forthcoming zuul job [2] (hosted by the plugin)
that will automate gabbi-based service testing as described
in https://anticdent.org/gabbi-in-the-gate.html .

By setting the environment in the devstack-tempest playbook
we avoid needing to duplicate the playbook: we can use
it directly and have less risk of plays diverging. The
calling job (which doesn't allow the "environment" key)
sets a "var" which then sets the environment variable.

[1] https://gabbi.readthedocs.io
[2] https://review.openstack.org/611678

Needed-By: https://review.openstack.org/611678
Change-Id: I5aa890baa489f434b95d21e4e618f5202ac550e4