Change way how second NIC in guest OS is configured
In case of RHEL based operating system used in guest vm in
test_network_v6 module, there is need to prepare

    /etc/sysconfig/network-scripts/ifcfg-%(nic)s

file with "IPV6INIT=yes" to enable IPv6 on it.
Otherwise IPv6 will not be configured on this NIC when:

    ip link set up dev %(nic)s

will be executed and tests with dualnet=False will fail.
This isn't necessary on systems like Ubuntu or Cirros.

Change-Id: I397b1c3584a045f9ca4f47b4e691385628d35efa
Closes-Bug: #1795891