Use remote_client from tempest.lib.common instead of tempest.common
This patch is replacing the test_attach_interfaces.py module to use the
RemoteClient from tempest.lib.common instead of tempest.common. Due to
the saner defaults from the new class tests such
"test_reassign_port_between_servers" is less prone to fail (see related
bug).

Also, this patch intentionally does not attempt to replace every
occurence of tempest.common.utils.linux.remote_client with its
counter-part in tempest.lib because that would require porting
some of the methods that exists in the old class to the new one.
The focus of this patch is to address the problem described in
the related bug.

Related-Bug: #1796882
Change-Id: I07bde710581270766f049c7147246971fed6cd91