Fix potential issue with adding a fixed ip
In test_add_remove_fixed_ip there is a potential issue where the user
running the test does not have the ability to create a server with a
fixed ip. As pointed out by zhufl (in review of change
I10a34fb12b595400fc15d935070f130c8e3d3963) This commit fixes the issue
so that we properly skip if tempest is not configured in a way where it
can create a server with the fixed ip.

Change-Id: Idcd66ac87cf156bc5c2b38178262c4686238bcd4