Make connectivity test less backend dependent
The test_connectivity_between_vms_on_different_networks should assert
connectivity between different VMS on different networks but the test is
also asserting that one VM can ping the services port (such as DHCP)
running on the other network and that may not be the case for all
backends.

OVN for example uses a "localport" type to implement a
DHCP port. That port type in OVN does not allow traffic
outside the network it belongs to [0] which is causing the test
test_connectivity_between_vms_on_different_networks to fail when used
with networking-ovn.

This patch is modifying the test to only ping the relevant ports when
asserting that the VMs have connectivity with each other.

[0]
https://github.com/openvswitch/ovs/blob/d9677a1f0eaf6352c44af3ba3a4c2faf76a2a39b/ovn/controller/physical.c#L1176-L1182

Closes-Bug: #1798317
Change-Id: Ia4a6ee65d3d737f8452b4aed9bf4624455a3b050