docs: Include information about the list of supported projects
This documentation is included in HACKING at the top to let
newcomers know that Tempest's testing scope is confined to the
following projects:

* Cinder
* Glance
* Keystone
* Neutron
* Nova
* Swift

The user is directed to [0] for additional Tempest plugin
repos where they can add their tests for other projects. Note
that [0] currently appears to not be rendering correctly,
but this is an unrelated issue.

[0] https://docs.openstack.org/tempest/latest/plugin-registry.html#detected-plugins

Change-Id: I800a3f6ac4c347d4322574f649cb052070127db6