Migrate tempest provided base jobs to bionic
All the base jobs provided by Tempest are migrated to run in bionic
from stein onwards. As Tempest is branchless, we need to keep running or
available those base job running on xenial for stein before which is
stable/rocky, stable/queens and stable/pike.

The selection of the correct nodeset for this happens mostly
automatically when we avoid selecting nodesets explicitly because we
inherit them from the correct branches of devstack which have the
matching job variants defined.

There are only two places left where we need to explicitly override the
default nodesets:

1. tempest-multinode-full
Here we need to select a nodeset with two nodes. Since this job runs
only on master, we simply select two Bionic nodes instead of two Xenial
nodes now.

2. Stable branch testing jobs like tempest-full-rocky
These do not use branch selectors, but "override-checkout" instead, so
they inherit Bionic nodesets from devstack master branch. We need to
force them to run on Xenial for backwards compatibility.

Drop the tempest-full-py36 job because it will be identical to the
tempest-full-py3 job now.

Depends-On: https://review.openstack.org/610977
Depends-On: https://review.openstack.org/611594
Change-Id: I5a24c697103090f6712d6e9168abe6c86f7c380c