Move test test_list_agents_non_admin to non-admin dir
This patch set moves the test test_list_agents_non_admin
out of the admin subdirectory because it is obviously
not an admin test.

This test is not referenced by interop or anything critical
so can be moved: http://codesearch.openstack.org/?q=test_list_agents_non_admin&i=nope&files=&repos=

Also changes test_delete_agent_negative to use
admin_agents_client as delete agent action requires
admin credentials:

https://github.com/openstack/neutron/blob/4b40a4c3231486898923cee70f9d70d6ad803769/etc/policy.json#L132

This is not essential for the test to pass but strictly speaking
the admin client should be used, not the non-admin client.

Change-Id: I0fca827be8109124a05fd5474715b766065abcf2