Immutable user source: Non-admin V2 and V3 tests
If the keystone user source is immutable, such as an LDAP
active directory implementation, tempest tests that try
to create or delete a user will fail. Instead of failing,
we would like them to skip. This change uses a testtools
decorator to avoid unnecessary modifications and allow those tests
to skip. In [1], I introduced the config setting that allows
this to happen.

[1] https://review.openstack.org/#/c/585536/

Change-Id: I786499204acdb929ee341e014bcb0459d2f27483
Partial-Bug: 1777047