Immutable user source: v3 test_projects
If the keystone user source is immutable, such as an LDAP
active directory implementation, tempest tests that try
to create or delete a user will fail. Instead of failing,
we would like them to skip. This change uses a testtools
decorator to avoid unnecessary modifications and allow those tests
to skip.

Change-Id: I3c4fa3578e12abb853bd427741b366d52017e3ad