Immutable user source: v3 test_groups
If the keystone user source is immutable, such as an LDAP
active directory implementation, tempest tests that try
to create or delete a user will fail. Instead of failing,
we would like them to skip. This change uses a testtools
decorator to avoid unnecessary modifications and allow those tests
to skip.

Partial-Bug: #1777047

Change-Id: I8a0d5935b695a1649b204bd3c3351f34a80e9059