immutable user source: v3 test_roles
If the keystone user source is immutable, such as an LDAP
active directory implementation, tempest tests that try
to create or delete a user will fail. Instead of failing,
we would like them to skip. This change uses a testtools
decorator to avoid unnecessary modifications and allow those tests
to skip.

Change-Id: Iba39c971468759d9b7b3f0382dfcb881cbb1801d