Mark test_server_connectivity_reboot as slow
Based on average test run times from the last 300 runs of
the tempest-full job [1] the test_server_connectivity_reboot
test is third slowest at about 200 seconds. This should not
be a surprise given all of the other tests in the same class
are already marked slow.

This change marks that test slow like the others which means
it will be run in the tempest-slow job.

[1] http://paste.openstack.org/show/745114/

Change-Id: Ie8ea1555cc4512bf29dff1e7df592dedfab28c61
Related-Bug: #1783405